+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# Unreleased
+
+Changed
+  * AZ-686: Allow sending other metrics when on query fails
+  * AZ-686: Bump requirements
+
+Fixed
+  * AZ-686: Fix authentication token scope for Log Analytics
+
 # v1.0.1 - 2022-01-14
 
 Changed